fix: merge config
